requirements.txt
Django~=3.2.6

python-decouple~=3.4
dj-database-url~=0.5.0
psycopg2-binary~=2.9.1
djangorestframework~=3.12.4

firebase-admin~=5.0.0

django-filter~=2.4.0
django-chartjs~=2.2.1
django-pivot~=1.8.1

drf-yasg==1.20.0


django-debug-toolbar~=3.2.1
gunicorn~=20.1.0